src/app/requirements.txt
db-dtypes>=1.0
google-cloud-bigquery>=3.0
streamlit>=1.40
